moved unsqueeze remapping into dim inference
